refactoring and more result structures
Refactor towards more use-cases. The VALUES structure
gets intermediate hash key layers so it can be extended
with more in the future. That should not hurt backcompat
as it wasn't returned yet. Now it is, though. The result
in wantarray context contains now also an array of the
single text lines instead of just a big blob, and also
that VALUES data structure.